careerChatbotBackend/requirements.txt
langchain-cli
langchain-community
langchain_google_genai
tavily-python
python-dotenv
fastapi
langchain_community
pypdf
uvicorn